Shared getParentUrl() function
