+++ src/main/java/org/craftercms/studio/impl/v1/util/ContentUtils.java
@@ -30,6 +30,9 @@ import java.io.*;
 import java.util.List;
 
 import static java.lang.String.format;
+import static org.apache.commons.io.FilenameUtils.getFullPathNoEndSeparator;
+import static org.apache.commons.lang3.StringUtils.removeEnd;
+import static org.craftercms.studio.api.v1.constant.DmConstants.SLASH_INDEX_FILE;
 import static org.craftercms.studio.api.v1.constant.StudioConstants.FILE_SEPARATOR;
 
 
@@ -98,9 +101,18 @@ public class ContentUtils {
 		return false;
 	}
 
-	public static String getParentUrl(String url) {
-		int lastIndex = url.lastIndexOf(FILE_SEPARATOR);
-		return url.substring(0, lastIndex);
+	/**
+	 * Get the parent url: for folders and components it's just parent, for pages it's the parent of the parent.
+	 * e.g.:
+	 * /site/website/articles/page1/index.xml -> /site/website/articles
+	 * /site/components/posts/january/clickbait.xml -> /site/components/posts/january
+	 * /site/components/articles/health/ -> /site/components/articles
+	 *
+	 * @param path path of the content item
+	 * @return path of the parent item
+	 */
+	public static String getParentUrl(String path) {
+		return getFullPathNoEndSeparator(removeEnd(path, SLASH_INDEX_FILE));
 	}
 
 	/**
