+++ src/test/java/org/craftercms/studio/impl/v2/service/clipboard/internal/ClipboardServiceInternalImplTest.java
@@ -43,12 +43,6 @@ import static org.mockito.Mockito.*;
 public class ClipboardServiceInternalImplTest {
 
 	private static final String SITE_ID = "mySite";
-	private static final String PAGE_URL = "/site/website/articles/testing/my-article/index.xml";
-	private static final String PAGE_URL_PARENT = "/site/website/articles/testing";
-	private static final String COMPONENT_URL = "/site/components/articles/testing/my-article.xml";
-	private static final String COMPONENT_URL_PARENT = "/site/components/articles/testing";
-	private static final String FOLDER_URL = "/site/components/articles/testing";
-	private static final String FOLDER_URL_PARENT = "/site/components/articles";
 
 	@Mock
 	private ItemService itemService;
@@ -303,24 +297,6 @@ public class ClipboardServiceInternalImplTest {
 	public void preventCutPastePageIntoSameFolder()
 			throws ServiceLayerException {
 		service.validatePasteItemsAction(SITE_ID, CUT, "/site/website/articles/article-1/index.xml", "/site/website/articles");
-	}
-
-	@Test
-	public void calculatePageParentUrlTest() {
-		String parentUrl = service.getParentUrl(PAGE_URL);
-		assertEquals(PAGE_URL_PARENT, parentUrl, format("Parent of '%s' does not match expected value", PAGE_URL));
-	}
-
-	@Test
-	public void calculateComponentPageParentUrlTest() {
-		String parentUrl = service.getParentUrl(COMPONENT_URL);
-		assertEquals(COMPONENT_URL_PARENT, parentUrl, format("Parent of '%s' does not match expected value", COMPONENT_URL));
-	}
-
-	@Test
-	public void calculateFolderParentUrlTest() {
-		String parentUrl = service.getParentUrl(FOLDER_URL);
-		assertEquals(FOLDER_URL_PARENT, parentUrl, format("Parent of '%s' does not match expected value", FOLDER_URL));
 	}
 
 	@Test
